requirements.txt
pillow==11.3.0
PyMuPDF==1.26.3
llama_cpp_python==0.3.14